# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "EvidenceLevel": {
        "description": "\u77e5\u8bc6\u6761\u76ee\u7684\u8bc1\u636e\u7b49\u7ea7\uff08docs/research/knowledge-base.md\uff09\u3002",
        "enum": [
          "E0",
          "E1",
          "E2",
          "E3",
          "E4"
        ],
        "title": "EvidenceLevel",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KnowledgeQuery": {
        "additionalProperties": false,
        "description": "\u4e00\u6b21\u77e5\u8bc6\u68c0\u7d22\u3002\n\n- `terms`\uff1a\u5168\u90e8\u987b\u51fa\u73b0\u5728\u6761\u76ee `name` / `claim` / `conditions` \u4e2d\uff08\u4e0d\u533a\u5206\u5927\u5c0f\u5199\uff0cAND \u8bed\u4e49\uff09\uff1b\n- `name_prefix`\uff1a\u5e93\u524d\u7f00\uff0c\u5982 `strategy_`\u3001`factor_`\u3001`state_`\uff08\u7a7a = \u4e0d\u9650\uff09\uff1b\n- `evidence_at_least`\uff1a\u53ea\u8fd4\u56de\u8bc1\u636e\u7b49\u7ea7\u4e0d\u4f4e\u4e8e\u8be5\u503c\u7684\u6761\u76ee\uff08\u7a7a = \u4e0d\u9650\uff09\uff1b\n- `statuses`\uff1a\u53ea\u8fd4\u56de\u8fd9\u4e9b\u72b6\u6001\uff08\u7a7a = \u4e0d\u9650\uff09\uff1b\n- `limit`\uff1a\u7ed3\u679c\u4e0a\u9650\u3002",
        "properties": {
          "evidence_at_least": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "limit": {
            "default": 50,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "name_prefix": {
            "default": "",
            "title": "Name Prefix",
            "type": "string"
          },
          "schema_version": {
            "default": "2.0.0",
            "pattern": "^(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)(?:-(?:0|[1-9][0-9]*|[0-9]*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9][0-9]*|[0-9]*[a-zA-Z-][0-9a-zA-Z-]*))*)?(?:\\+[0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*)?$",
            "title": "Schema Version",
            "type": "string"
          },
          "statuses": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/KnowledgeStatus"
            },
            "title": "Statuses",
            "type": "array"
          },
          "terms": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Terms",
            "type": "array"
          }
        },
        "title": "KnowledgeQuery",
        "type": "object"
      },
      "KnowledgeStatus": {
        "enum": [
          "unverified",
          "supported",
          "contradicted",
          "inconclusive"
        ],
        "title": "KnowledgeStatus",
        "type": "string"
      },
      "ReportEnvelope": {
        "description": "The minimal envelope every report kind is served through.",
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ReportKind"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "kind",
          "id",
          "created",
          "payload",
          "content_hash"
        ],
        "title": "ReportEnvelope",
        "type": "object"
      },
      "ReportKind": {
        "description": "The report kinds the console serves (task spec: (a)-(d)).",
        "enum": [
          "validation_report",
          "research_loop_round",
          "state_strategy_matrix",
          "router_paper_run"
        ],
        "title": "ReportKind",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "HLENS-AutoResearch API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/contracts": {
      "get": {
        "operationId": "contracts_contracts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Contracts Contracts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Contracts"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/knowledge/search": {
      "post": {
        "operationId": "knowledge_search_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Search"
      }
    },
    "/lifecycle/transitions": {
      "get": {
        "operationId": "transitions_lifecycle_transitions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "title": "Response Transitions Lifecycle Transitions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Transitions"
      }
    },
    "/reports/{kind}": {
      "get": {
        "operationId": "list_reports_reports__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReportKind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReportEnvelope"
                  },
                  "title": "Response List Reports Reports  Kind  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reports"
      }
    },
    "/reports/{kind}/{report_id}": {
      "get": {
        "operationId": "get_report_reports__kind___report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReportKind"
            }
          },
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report"
      }
    }
  }
}